+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 - support client roles
+- add common claims that Keycloak provides by default (`iss`, `aud`, `iat`, `jti` and `azp`)
+- change the type of the `sub` claim from `String` to `Uuid`
 
 ## 0.1.0